Fixing a bug in timeslides function Original: b21e4be2319f7049a1c5c9474191c4d9e3dc43ff

+++ lal/packages/tools/include/LIGOMetadataUtils.h
@@ -763,7 +763,7 @@ XLALTimeSlideSnglBurst (
 	SnglBurstTable *triggerlist,
 	INT8 startTime,
 	INT8 stopTime,
-	int slideTime
+	INT8 slideTime
 );
 
 #if 0
